requirements.txt
av
pyocr
Pillow
